modificados los desgastes y sus tests

+++ proyecto/src/DesgastePorPorcentaje.java
@@ -7,7 +7,7 @@ public class DesgastePorPorcentaje extends Desgaste {
     }
 
     @Override
-    public double aplicar(int durabilidad, int fuerza){
+    public int aplicar(int durabilidad, int fuerza){
         durabilidad -= (fuerza / porcentaje);
 
         return durabilidad;
